[frontend] refacto widgets private dashboard for entities
entities tree
entities heatmap
entities radar
entities donut
entities timeline
entities multi area
entities multiline
entities vertical bar
entities distrib list
entities list
entities bookmarks
