+++ opencti-platform/opencti-front/src/components/dashboard/WidgetDistributionList.tsx
@@ -1,0 +1,102 @@
+import List from '@mui/material/List';
+import ListItem from '@mui/material/ListItem';
+import { Link } from 'react-router-dom';
+import ListItemIcon from '@mui/material/ListItemIcon';
+import ListItemText from '@mui/material/ListItemText';
+import React from 'react';
+import { useTheme } from '@mui/styles';
+import ItemIcon from '../ItemIcon';
+import { computeLink } from '../../utils/Entity';
+import type { Theme } from '../Theme';
+import { useFormatter } from '../i18n';
+
+interface WidgetDistributionListProps {
+  data: any[]
+  hasSettingAccess?: boolean
+}
+
+const WidgetDistributionList = ({
+  data,
+  hasSettingAccess = false,
+}: WidgetDistributionListProps) => {
+  const theme = useTheme<Theme>();
+  const { n } = useFormatter();
+
+  return (
+    <div
+      id="container"
+      style={{
+        width: '100%',
+        height: '100%',
+        overflow: 'auto',
+        paddingBottom: 10,
+        marginBottom: 10,
+      }}
+    >
+      <List style={{ marginTop: -10 }}>
+        {data.map((entry) => {
+          let link: string | null = null;
+          if (entry.type !== 'User' || hasSettingAccess) {
+            link = entry.id ? computeLink(entry) : null;
+          }
+          return (
+            <ListItem
+              key={entry.label}
+              dense={true}
+              button={!!link}
+              divider={true}
+              component={link ? Link : null}
+              to={link || null}
+              sx={{
+                height: 50,
+                minHeight: 50,
+                maxHeight: 50,
+                paddingRight: 0,
+              }}
+            >
+              <ListItemIcon>
+                <ItemIcon
+                  color={
+                    theme.palette.mode === 'light'
+                    && entry.color === '#ffffff'
+                      ? '#000000'
+                      : entry.color
+                  }
+                  type={entry.type}
+                />
+              </ListItemIcon>
+              <ListItemText
+                primary={
+                  <div
+                    style={{
+                      whiteSpace: 'nowrap',
+                      overflow: 'hidden',
+                      textOverflow: 'ellipsis',
+                      paddingRight: 10,
+                    }}
+                  >
+                    {entry.label}
+                  </div>
+                }
+              />
+              <div
+                style={{
+                  float: 'right',
+                  marginRight: 20,
+                  fontSize: 20,
+                  whiteSpace: 'nowrap',
+                  overflow: 'hidden',
+                  textOverflow: 'ellipsis',
+                }}
+              >
+                {n(entry.value)}
+              </div>
+            </ListItem>
+          );
+        })}
+      </List>
+    </div>
+  );
+};
+
+export default WidgetDistributionList;
