+++ opencti-platform/opencti-front/src/private/components/common/stix_core_objects/StixCoreObjectsDistributionList.jsx
@@ -1,60 +1,13 @@
 import React from 'react';
 import { graphql } from 'react-relay';
-import CircularProgress from '@mui/material/CircularProgress';
-import Paper from '@mui/material/Paper';
-import Typography from '@mui/material/Typography';
-import makeStyles from '@mui/styles/makeStyles';
-import List from '@mui/material/List';
-import ListItem from '@mui/material/ListItem';
-import { Link } from 'react-router-dom';
-import ListItemIcon from '@mui/material/ListItemIcon';
-import ListItemText from '@mui/material/ListItemText';
-import { useTheme } from '@mui/styles';
 import { QueryRenderer } from '../../../../relay/environment';
 import { useFormatter } from '../../../../components/i18n';
-import { resolveLink } from '../../../../utils/Entity';
-import ItemIcon from '../../../../components/ItemIcon';
 import useGranted, { SETTINGS_SETACCESSES } from '../../../../utils/hooks/useGranted';
 import { buildFiltersAndOptionsForWidgets } from '../../../../utils/filters/filtersUtils';
-
-const useStyles = makeStyles({
-  container: {
-    width: '100%',
-    height: '100%',
-    overflow: 'auto',
-    paddingBottom: 10,
-    marginBottom: 10,
-  },
-  paper: {
-    height: '100%',
-    margin: '10px 0 0 0',
-    padding: 0,
-    borderRadius: 4,
-  },
-  item: {
-    height: 50,
-    minHeight: 50,
-    maxHeight: 50,
-    paddingRight: 0,
-  },
-  itemText: {
-    whiteSpace: 'nowrap',
-    overflow: 'hidden',
-    textOverflow: 'ellipsis',
-    paddingRight: 10,
-  },
-});
-
-const inlineStyles = {
-  itemNumber: {
-    float: 'right',
-    marginRight: 20,
-    fontSize: 20,
-    whiteSpace: 'nowrap',
-    overflow: 'hidden',
-    textOverflow: 'ellipsis',
-  },
-};
+import WidgetContainer from '../../../../components/dashboard/WidgetContainer';
+import WidgetNoData from '../../../../components/dashboard/WidgetNoData';
+import WidgetLoader from '../../../../components/dashboard/WidgetLoader';
+import WidgetDistributionList from '../../../../components/dashboard/WidgetDistributionList';
 
 const stixCoreObjectsDistributionListDistributionQuery = graphql`
   query StixCoreObjectsDistributionListDistributionQuery(
@@ -137,22 +90,13 @@ const StixCoreObjectsDistributionList = ({
   dataSelection,
   parameters = {},
 }) => {
-  const classes = useStyles();
-  const theme = useTheme();
-  const { t_i18n, n } = useFormatter();
+  const { t_i18n } = useFormatter();
   const hasSetAccess = useGranted([SETTINGS_SETACCESSES]);
   const renderContent = () => {
     const selection = dataSelection[0];
     const dataSelectionTypes = ['Stix-Core-Object'];
     const { filters, dataSelectionElementId, dataSelectionToTypes } = buildFiltersAndOptionsForWidgets(selection.filters);
-    const computeLink = (entry) => {
-      const resolution = resolveLink(entry.type);
-      if (resolution) {
-        // Handle type with no link
-        return `${resolution}/${entry.id}`;
-      }
-      return null;
-    };
+
     return (
       <QueryRenderer
         query={stixCoreObjectsDistributionListDistributionQuery}
@@ -190,108 +134,24 @@ const StixCoreObjectsDistributionList = ({
               id: selection.attribute.endsWith('_id') ? o.entity.id : null,
               type: o.entity?.entity_type ?? o.label,
             }));
-            return (
-              <div id="container" className={classes.container}>
-                <List style={{ marginTop: -10 }}>
-                  {data.map((entry) => {
-                    // eslint-disable-next-line no-nested-ternary
-                    const link = entry.type === 'User' && !hasSetAccess
-                      ? null
-                      : entry.id
-                        ? computeLink(entry)
-                        : null;
-                    return (
-                      <ListItem
-                        key={entry.label}
-                        dense={true}
-                        button={!!link}
-                        classes={{ root: classes.item }}
-                        divider={true}
-                        component={link ? Link : null}
-                        to={link || null}
-                      >
-                        <ListItemIcon>
-                          <ItemIcon
-                            color={
-                              theme.palette.mode === 'light'
-                              && entry.color === '#ffffff'
-                                ? '#000000'
-                                : entry.color
-                            }
-                            type={entry.type}
-                          />
-                        </ListItemIcon>
-                        <ListItemText
-                          primary={
-                            <div className={classes.itemText}>
-                              {entry.label}
-                            </div>
-                          }
-                        />
-                        <div style={inlineStyles.itemNumber}>
-                          {n(entry.value)}
-                        </div>
-                      </ListItem>
-                    );
-                  })}
-                </List>
-              </div>
-            );
+            return <WidgetDistributionList data={data} hasSettingAccess={hasSetAccess} />;
           }
           if (props) {
-            return (
-              <div style={{ display: 'table', height: '100%', width: '100%' }}>
-                <span
-                  style={{
-                    display: 'table-cell',
-                    verticalAlign: 'middle',
-                    textAlign: 'center',
-                  }}
-                >
-                  {t_i18n('No entities of this type has been found.')}
-                </span>
-              </div>
-            );
+            return <WidgetNoData />;
           }
-          return (
-            <div style={{ display: 'table', height: '100%', width: '100%' }}>
-              <span
-                style={{
-                  display: 'table-cell',
-                  verticalAlign: 'middle',
-                  textAlign: 'center',
-                }}
-              >
-                <CircularProgress size={40} thickness={2} />
-              </span>
-            </div>
-          );
+          return <WidgetLoader />;
         }}
       />
     );
   };
   return (
-    <div style={{ height: height || '100%' }}>
-      <Typography
-        variant="h4"
-        gutterBottom={true}
-        style={{
-          margin: variant !== 'inLine' ? '0 0 10px 0' : '-10px 0 10px -7px',
-          whiteSpace: 'nowrap',
-          overflow: 'hidden',
-          textOverflow: 'ellipsis',
-        }}
-      >
-        {parameters.title || t_i18n('Distribution of entities')}
-      </Typography>
-      {variant === 'inLine' ? (
-        renderContent()
-      ) : (
-        <Paper classes={{ root: classes.paper }} variant="outlined">
-          {renderContent()}
-        </Paper>
-      )}
-    </div>
+    <WidgetContainer
+      height={height}
+      title={parameters.title ?? t_i18n('Distribution of entities')}
+      variant={variant}
+    >
+      {renderContent()}
+    </WidgetContainer>
   );
 };
