+++ opencti-platform/opencti-front/src/private/components/common/stix_core_objects/StixCoreObjectsMultiHeatMap.jsx
@@ -1,54 +1,13 @@
 import React from 'react';
 import { graphql } from 'react-relay';
-import CircularProgress from '@mui/material/CircularProgress';
-import Paper from '@mui/material/Paper';
-import Typography from '@mui/material/Typography';
-import makeStyles from '@mui/styles/makeStyles';
-import { useTheme } from '@mui/styles';
-import Chart from '../charts/Chart';
 import { QueryRenderer } from '../../../../relay/environment';
 import { useFormatter } from '../../../../components/i18n';
 import { monthsAgo, now } from '../../../../utils/Time';
-import { heatMapOptions } from '../../../../utils/Charts';
 import { removeEntityTypeAllFromFilterGroup } from '../../../../utils/filters/filtersUtils';
-
-const useStyles = makeStyles(() => ({
-  paper: {
-    minHeight: 280,
-    height: '100%',
-    margin: '4px 0 0 0',
-    padding: '0 0 10px 0',
-    borderRadius: 4,
-  },
-}));
-
-const darkColors = [
-  '#001e3c',
-  '#023362',
-  '#02407a',
-  '#045198',
-  '#0561b4',
-  '#0b75d9',
-  '#2986e7',
-  '#3a95f3',
-  '#4da3ff',
-  '#76bbff',
-  '#9eccff',
-];
-
-const lightColors = [
-  '#f3f6f9',
-  '#76bbff',
-  '#4da3ff',
-  '#3a95f3',
-  '#2986e7',
-  '#0b75d9',
-  '#0561b4',
-  '#02407a',
-  '#023362',
-  '#001e3c',
-  '#021428',
-];
+import WidgetNoData from '../../../../components/dashboard/WidgetNoData';
+import WidgetLoader from '../../../../components/dashboard/WidgetLoader';
+import WidgetContainer from '../../../../components/dashboard/WidgetContainer';
+import WidgetMultiHeatMap from '../../../../components/dashboard/WidgetMultiHeatMap';
 
 const stixCoreObjectsMultiHeatMapTimeSeriesQuery = graphql`
   query StixCoreObjectsMultiHeatMapTimeSeriesQuery(
@@ -81,9 +40,7 @@ const StixCoreObjectsMultiHeatMap = ({
   withExportPopover = false,
   isReadOnly = false,
 }) => {
-  const theme = useTheme();
-  const classes = useStyles();
-  const { t_i18n, fsd } = useFormatter();
+  const { t_i18n } = useFormatter();
   const renderContent = () => {
     const timeSeriesParameters = dataSelection.map((selection) => {
       return {
@@ -122,100 +79,34 @@ const StixCoreObjectsMultiHeatMap = ({
               .flat();
             const maxValue = Math.max(...allValues);
             const minValue = Math.min(...allValues);
-            const interval = Math.trunc((maxValue - minValue) / 9);
-            const colorRanges = Array(10)
-              .fill(0)
-              .map((_, i) => ({
-                from:
-                  minValue + (i + 1) * interval - interval === 0
-                    ? 1
-                    : minValue + (i + 1) * interval - interval,
-                to: minValue + (i + 1) * interval,
-                color:
-                  theme.palette.mode === 'dark'
-                    ? darkColors[i + 1]
-                    : lightColors[i + 1],
-              }));
-            colorRanges.push({
-              from: 0,
-              to: 0,
-              color:
-                theme.palette.mode === 'dark' ? darkColors[0] : lightColors[0],
-            });
+
             return (
-              <Chart
-                options={heatMapOptions(
-                  theme,
-                  true,
-                  fsd,
-                  undefined,
-                  undefined,
-                  parameters.stacked,
-                  colorRanges,
-                )}
-                series={chartData}
-                type="heatmap"
-                width="100%"
-                height="100%"
-                withExportPopover={withExportPopover}
-                isReadOnly={isReadOnly}
+              <WidgetMultiHeatMap
+                data={chartData}
+                minValue={minValue}
+                maxValue={maxValue}
+                isStacked={parameters.stacked}
+                withExport={withExportPopover}
+                readonly={isReadOnly}
               />
             );
           }
           if (props) {
-            return (
-              <div style={{ display: 'table', height: '100%', width: '100%' }}>
-                <span
-                  style={{
-                    display: 'table-cell',
-                    verticalAlign: 'middle',
-                    textAlign: 'center',
-                  }}
-                >
-                  {t_i18n('No entities of this type has been found.')}
-                </span>
-              </div>
-            );
+            return <WidgetNoData />;
           }
-          return (
-            <div style={{ display: 'table', height: '100%', width: '100%' }}>
-              <span
-                style={{
-                  display: 'table-cell',
-                  verticalAlign: 'middle',
-                  textAlign: 'center',
-                }}
-              >
-                <CircularProgress size={40} thickness={2} />
-              </span>
-            </div>
-          );
+          return <WidgetLoader />;
         }}
       />
     );
   };
   return (
-    <div style={{ height: height || '100%' }}>
-      <Typography
-        variant="h4"
-        gutterBottom={true}
-        style={{
-          margin: variant !== 'inLine' ? '0 0 10px 0' : '-10px 0 10px -7px',
-          whiteSpace: 'nowrap',
-          overflow: 'hidden',
-          textOverflow: 'ellipsis',
-        }}
-      >
-        {parameters.title ?? t_i18n('Entities history')}
-      </Typography>
-      {variant !== 'inLine' ? (
-        <Paper classes={{ root: classes.paper }} variant="outlined">
-          {renderContent()}
-        </Paper>
-      ) : (
-        renderContent()
-      )}
-    </div>
+    <WidgetContainer
+      height={height}
+      title={parameters.title ?? t_i18n('Entities history')}
+      variant={variant}
+    >
+      {renderContent()}
+    </WidgetContainer>
   );
 };
