+++ opencti-platform/opencti-front/src/private/components/common/stix_core_objects/StixCoreObjectsTreeMap.jsx
@@ -1,25 +1,12 @@
 import React from 'react';
 import { graphql } from 'react-relay';
-import CircularProgress from '@mui/material/CircularProgress';
-import Paper from '@mui/material/Paper';
-import Typography from '@mui/material/Typography';
-import makeStyles from '@mui/styles/makeStyles';
-import { useTheme } from '@mui/styles';
-import Chart from '../charts/Chart';
 import { QueryRenderer } from '../../../../relay/environment';
 import { useFormatter } from '../../../../components/i18n';
-import { treeMapOptions } from '../../../../utils/Charts';
-import { defaultValue } from '../../../../utils/Graph';
 import { buildFiltersAndOptionsForWidgets } from '../../../../utils/filters/filtersUtils';
-
-const useStyles = makeStyles(() => ({
-  paper: {
-    height: '100%',
-    margin: '10px 0 0 0',
-    padding: 0,
-    borderRadius: 4,
-  },
-}));
+import WidgetNoData from '../../../../components/dashboard/WidgetNoData';
+import WidgetLoader from '../../../../components/dashboard/WidgetLoader';
+import WidgetContainer from '../../../../components/dashboard/WidgetContainer';
+import WidgetTree from '../../../../components/dashboard/WidgetTree';
 
 const stixCoreObjectsTreeMapDistributionQuery = graphql`
   query StixCoreObjectsTreeMapDistributionQuery(
@@ -205,8 +192,6 @@ const StixCoreObjectsTreeMap = ({
   withExportPopover = false,
   isReadOnly = false,
 }) => {
-  const classes = useStyles();
-  const theme = useTheme();
   const { t_i18n } = useFormatter();
   const renderContent = () => {
     const selection = dataSelection[0];
@@ -237,83 +222,32 @@ const StixCoreObjectsTreeMap = ({
             && props.stixCoreObjectsDistribution.length > 0
           ) {
             const data = props.stixCoreObjectsDistribution;
-            const chartData = data.map((n) => ({
-              // eslint-disable-next-line no-nested-ternary
-              x: selection.attribute.endsWith('_id')
-                ? defaultValue(n.entity)
-                : selection.attribute === 'entity_type'
-                  ? t_i18n(`entity_${n.label}`)
-                  : n.label,
-              y: n.value,
-            }));
-            const series = [{ data: chartData }];
             return (
-              <Chart
-                options={treeMapOptions(
-                  theme,
-                  'bottom',
-                  parameters.distributed,
-                )}
-                series={series}
-                type="treemap"
-                width="100%"
-                height="100%"
-                withExportPopover={withExportPopover}
-                isReadOnly={isReadOnly}
+              <WidgetTree
+                data={data}
+                groupBy={selection.attribute}
+                isDistributed={parameters.distributed}
+                readonly={isReadOnly}
+                withExport={withExportPopover}
               />
             );
           }
           if (props) {
-            return (
-              <div style={{ display: 'table', height: '100%', width: '100%' }}>
-                <span
-                  style={{
-                    display: 'table-cell',
-                    verticalAlign: 'middle',
-                    textAlign: 'center',
-                  }}
-                >
-                  {t_i18n('No entities of this type has been found.')}
-                </span>
-              </div>
-            );
+            return <WidgetNoData />;
           }
-          return (
-            <div style={{ display: 'table', height: '100%', width: '100%' }}>
-              <span
-                style={{
-                  display: 'table-cell',
-                  verticalAlign: 'middle',
-                  textAlign: 'center',
-                }}
-              >
-                <CircularProgress size={40} thickness={2} />
-              </span>
-            </div>
-          );
+          return <WidgetLoader />;
         }}
       />
     );
   };
   return (
-    <div style={{ height: height || '100%' }}>
-      <Typography
-        variant="h4"
-        gutterBottom={true}
-        style={{
-          margin: variant !== 'inLine' ? '0 0 10px 0' : '-10px 0 10px -7px',
-        }}
-      >
-        {parameters.title || t_i18n('Distribution of entities')}
-      </Typography>
-      {variant === 'inLine' ? (
-        renderContent()
-      ) : (
-        <Paper classes={{ root: classes.paper }} variant="outlined">
-          {renderContent()}
-        </Paper>
-      )}
-    </div>
+    <WidgetContainer
+      height={height}
+      title={parameters.title ?? t_i18n('Distribution of entities')}
+      variant={variant}
+    >
+      {renderContent()}
+    </WidgetContainer>
   );
 };
